+++ src/Bankly.php
@@ -5,6 +5,13 @@ namespace WeDevBr\Bankly;
 use Illuminate\Http\Client\RequestException;
 use Illuminate\Support\Facades\Http;
 use Ramsey\Uuid\Uuid;
+use TypeError;
+use WeDevBr\Bankly\Inputs\Customer;
+use WeDevBr\Bankly\Inputs\DocumentAnalysis;
+use WeDevBr\Bankly\Support\Contracts\CustomerInterface;
+use WeDevBr\Bankly\Support\Contracts\DocumentInterface;
+use WeDevBr\Bankly\Types\Pix\PixEntries;
+use WeDevBr\Bankly\Types\VirtualCard\VirtualCard;
 
 /**
  * Class Bankly
@@ -20,17 +27,19 @@ class Bankly
     private $token_expiry = 0;
     private $token = null;
     private $api_version = '1.0';
+    private $headers;
 
     /**
      * Bankly constructor.
-     * @param string $client_secret provided by Bankly Staff
-     * @param string $client_id provided by Bankly Staff
+     * @param null|string $client_secret provided by Bankly Staff
+     * @param null|string $client_id provided by Bankly Staff
      */
     public function __construct($client_secret = null, $client_id = null)
     {
         $this->api_url = config('bankly')['api_url'];
         $this->login_url = config('bankly')['login_url'];
         $this->setClientCredentials(['client_secret' => $client_secret, 'client_id' => $client_id]);
+        $this->headers = ['API-Version' => $this->api_version];
     }
 
     /**
@@ -48,22 +57,39 @@ class Bankly
      * @return array|mixed
      * @throws RequestException
      */
-    final public function getBankList()
+    public function getBankList()
     {
         return $this->get('/banklist');
     }
 
     /**
+     * Retrieve your balance account
      * @param string $branch
      * @param string $account
      * @return array|mixed
      * @throws RequestException
+     * @note If you have a RequestException on this endpoint in staging environment, please use getAccount() method instead.
      */
-    final public function getBalance(string $branch, string $account)
+    public function getBalance(string $branch, string $account)
+
     {
         return $this->get('/account/balance', [
             'branch' => $branch,
             'account' => $account
+        ]);
+    }
+
+    /**
+     * @param string $account
+     * @param string $includeBalance
+     * @return array|mixed
+     * @throws RequestException
+     * @note This method on this date (2020-10-21) works only on staging environment. Contact Bankly/Acesso for more details
+     */
+    public function getAccount(string $account, string $includeBalance = 'true')
+    {
+        return $this->get('/accounts/' . $account, [
+            'includeBalance' => $includeBalance,
         ]);
     }
 
@@ -77,21 +103,21 @@ class Bankly
      * @return array|mixed
      * @throws RequestException
      */
-    final public function getStatement(
+    public function getStatement(
         $branch,
         $account,
         $offset = 1,
         $limit = 20,
-        $details = 'true',
-        $detailsLevelBasic = 'true'
+        string $details = 'true',
+        string $detailsLevelBasic = 'true'
     ) {
         return $this->get('/account/statement', array(
             'branch' => $branch,
             'account' => $account,
             'offset' => $offset,
             'limit' => $limit,
-            'details' => (string) $details,
-            'detailsLevelBasic' => (string) $detailsLevelBasic
+            'details' => $details,
+            'detailsLevelBasic' => $detailsLevelBasic
         ));
     }
 
@@ -103,22 +129,27 @@ class Bankly
      * @param string $include_details
      * @return array|mixed
      * @throws RequestException
+     * @note This endpoint has been deprecated for some clients.
+     * You need to check with Acesso/Bankly if your environment has different parameters also.
+     * The response of this request does not have a default interface between environments.
+     * Pay attention when use this in your project.
      */
     public function getEvents(
         string $branch,
         string $account,
         int $page = 1,
         int $pagesize = 20,
-        $include_details = 'true'
+        string $include_details = 'true'
     ) {
         return $this->get(
             '/events',
             [
-                'Branch' => $branch,
-                'Account' => $account,
-                'Page' => $page,
-                'Pagesize' => $pagesize,
-                'IncludeDetails' => (string) $include_details
+                'branch' => $branch,
+                'account' => $account,
+                'page' => $page,
+                'pageSize' => $pagesize,
+                'includeDetails' => $include_details
+
             ]
         );
     }
@@ -157,6 +188,45 @@ class Bankly
     }
 
     /**
+     * Get transfer funds from an account
+     * @param string $branch
+     * @param string $account
+     * @param int $pageSize
+     * @param string|null $nextPage
+     * @return array|mixed
+     * @throws RequestException
+     */
+    public function getTransferFunds(string $branch, string $account, int $pageSize = 10, string $nextPage = null)
+    {
+        $queryParams = [
+            'branch' => $branch,
+            'account' => $account,
+            'pageSize' => $pageSize
+        ];
+        if ($nextPage) {
+            $queryParams['nextPage'] = $nextPage;
+        }
+        return $this->get('/fund-transfers', $queryParams);
+    }
+
+    /**
+     * Get Transfer Funds By Authentication Code
+     * @param string $branch
+     * @param string $account
+     * @param string $authenticationCode
+     * @return array|mixed
+     * @throws RequestException
+     */
+    public function findTransferFundByAuthCode(string $branch, string $account, string $authenticationCode)
+    {
+        $queryParams = [
+            'branch' => $branch,
+            'account' => $account
+        ];
+        return $this->get('/fund-transfers/' . $authenticationCode, $queryParams);
+    }
+
+    /**
      * @param string $branch
      * @param string $account
      * @param string $authentication_id
@@ -172,13 +242,164 @@ class Bankly
     }
 
     /**
+     * @param string $documentNumber
+     * @param DocumentAnalysis $document
+     * @param string $correlationId
+     * @return array|mixed
+     * @throws RequestException
+     */
+    public function documentAnalysis(
+        string $documentNumber,
+        $document,
+        string $correlationId = null
+    ) {
+        if (!$document instanceof DocumentInterface) {
+            throw new TypeError('The document must be an instance of DocumentInterface');
+        }
+
+        return $this->put(
+            "/document-analysis/{$documentNumber}",
+            [
+                'documentType' => $document->getDocumentType(),
+                'documentSide' => $document->getDocumentSide(),
+            ],
+            $correlationId,
+            true,
+            true,
+            $document
+        );
+    }
+
+    /**
+     * @param string $documentNumber
+     * @param array $tokens
+     * @param string $resultLevel
+     * @param string $correlationId
+     * @return array|mixed
+     */
+    public function getDocumentAnalysis(
+        string $documentNumber,
+        array $tokens = [],
+        string $resultLevel = 'ONLY_STATUS',
+        string $correlationId = null
+    ) {
+        $query = collect($tokens)
+            ->map(function ($token) {
+                return "token={$token}";
+            })
+            ->concat(["resultLevel={$resultLevel}"])
+            ->implode('&');
+
+        return $this->get(
+            "/document-analysis/{$documentNumber}",
+            $query,
+            $correlationId
+        );
+    }
+
+    /**
+     * @param string $documentNumber
+     * @param Customer $customer
+     * @param string $correlationId
+     * @return array|mixed
+     * @throws RequestException
+     */
+    public function customer(
+        string $documentNumber,
+        $customer,
+        string $correlationId = null
+    ) {
+        if (!$customer instanceof CustomerInterface) {
+            throw new TypeError('The customer must be an instance of CustomerInterface');
+        }
+
+        return $this->put("/customers/{$documentNumber}", $customer->toArray(), $correlationId);
+    }
+
+    /**
+     * Validate of boleto or dealership
+     *
+     * @param string $code - Digitable line
+     * @param string $correlationId
+     * @return array|mixed
+     * @throws RequestException
+     */
+    public function paymentValidate(string $code, string $correlationId)
+    {
+        return $this->post('/bill-payment/validate', ['code' => $code], $correlationId, true);
+    }
+
+    /**
+     * Confirmation of payment of boleto or dealership
+     *
+     * @param BillPayment $billPayment
+     * @param string $correlationId
+     * @return array|mixed
+     */
+    public function paymentConfirm(
+        BillPayment $billPayment,
+        string $correlationId
+    ) {
+        return $this->post('/bill-payment/confirm', $billPayment->toArray(), $correlationId, true);
+    }
+
+    /**
+     * Create a new PIX key link with account.
+     *
+     * @param PixEntries $pixEntries
+     * @return array|mixed
+     */
+    public function registerPixKey(PixEntries $pixEntries)
+    {
+        return $this->post('/pix/entries', [
+            'addressingKey' => $pixEntries->addressingKey->toArray(),
+            'account' => $pixEntries->account->toArray(),
+        ], null, true);
+    }
+
+    /**
+     * Gets the list of address keys linked to an account.
+     *
+     * @param string $accountNumber
+     * @return array|mixed
+     */
+    public function getPixAddressingKeys(string $accountNumber)
+    {
+        return $this->get("/accounts/$accountNumber/addressing-keys");
+    }
+
+    /**
+     * Gets details of the account linked to an addressing key.
+     *
+     * @param string $documentNumber
+     * @param string $addressinKeyValue
+     * @return array|mixed
+     */
+    public function getPixAddressingKeyValue(string $documentNumber, string $addressinKeyValue)
+    {
+        $this->setHeaders(['x-bkly-pix-user-id' => $documentNumber]);
+        return $this->get("/pix/entries/$addressinKeyValue");
+    }
+
+    /**
+     * Delete a key link with account.
+     *
+     * @param string $addressingKeyValue
+     * @return array|mixed
+     */
+    public function deletePixAddressingKeyValue(string $addressingKeyValue)
+    {
+        return $this->delete("/pix/entries/$addressingKeyValue");
+    }
+
+    /**
      * @param string $endpoint
-     * @param array $query
+     * @param array|string|null $query
      * @param null $correlation_id
      * @return array|mixed
      * @throws RequestException
      */
-    private function get(string $endpoint, array $query = null, $correlation_id = null)
+    private function get(string $endpoint, $query = null, $correlation_id = null)
     {
         if (now()->unix() > $this->token_expiry || !$this->token) {
             $this->auth();
@@ -193,6 +414,18 @@ class Bankly
             ->get($this->getFinalUrl($endpoint), $query)
             ->throw()
             ->json();
+    }
+
+    /**
+     * Create a new virtual card
+     *
+     * @param VirtualCard $virtualCard
+     * @return array|mixed
+     * @throws RequestException
+     */
+    public function virtualCard(VirtualCard $virtualCard)
+    {
+        return $this->post('/cards/virtual', $virtualCard->toArray(), null, true);
     }
 
     /**
@@ -225,6 +458,70 @@ class Bankly
     }
 
     /**
+     * @param string $endpoint
+     * @param array|null $body
+     * @param string|null $correlation_id
+     * @param bool $asJson
+     * @param bool $attachment
+     * @param DocumentAnalysis $document
+     * @param string $fieldName
+     * @return array|mixed
+     * @throws RequestException
+     */
+    private function put(
+        string $endpoint,
+        array $body = [],
+        string $correlation_id = null,
+        bool $asJson = false,
+        bool $attachment = false,
+        DocumentAnalysis $document = null
+    ) {
+        if (now()->unix() > $this->token_expiry || !$this->token) {
+            $this->auth();
+        }
+
+        if (is_null($correlation_id) && $this->requireCorrelationId($endpoint)) {
+            $correlation_id = Uuid::uuid4()->toString();
+        }
+
+        $body_format = $asJson ? 'json' : 'form_params';
+
+        $request = Http
+            ::withToken($this->token)
+            ->withHeaders($this->getHeaders(['x-correlation-id' => $correlation_id]))
+            ->bodyFormat($body_format);
+
+        if ($attachment) {
+            $request->attach($document->getFieldName(), $document->getFileContents(), $document->getFileName());
+        }
+
+        return $request->put($this->getFinalUrl($endpoint), $body)
+            ->throw()
+            ->json();
+    }
+
+    /**
+     * Http delete method.
+     *
+     * @param string $endpoint
+     * @return array|mixed
+     * @throws RequestException
+     */
+    private function delete(string $endpoint)
+    {
+        if (now()->unix() > $this->token_expiry || !$this->token) {
+            $this->auth();
+        }
+
+        $request = Http::withToken($this->token)
+            ->withHeaders($this->getHeaders($this->headers));
+
+        return $request->delete($this->getFinalUrl($endpoint))
+            ->throw()
+            ->json();
+    }
+
+    /**
      * @param string $version API version
      * @return $this
      */
@@ -238,11 +535,9 @@ class Bankly
      * @param array $headers
      * @return array|string[]
      */
-    final private function getHeaders($headers = [])
+    private function getHeaders($headers = [])
     {
-        $default_headers = [
-            'API-Version' => $this->api_version
-        ];
+        $default_headers = $this->headers;
 
         if (count($headers) > 0) {
             $default_headers = array_merge($headers, $default_headers);
@@ -252,10 +547,19 @@ class Bankly
     }
 
     /**
+     * @param array $header
+     * @return void
+     */
+    private function setHeaders($header)
+    {
+        $this->headers = array_merge($this->headers, $header);
+    }
+
+    /**
      * @param string $endpoint
      * @return bool
      */
-    final private function requireCorrelationId($endpoint)
+    private function requireCorrelationId(string $endpoint)
     {
         $not_required_endpoints = [
             '/banklist',
@@ -269,7 +573,7 @@ class Bankly
      * @param string $endpoint
      * @return string
      */
-    final private function getFinalUrl($endpoint)
+    private function getFinalUrl(string $endpoint)
     {
         return $this->api_url . $endpoint;
     }