+++ tests/BanklyTest.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * BanklyTest class
  *
@@ -90,6 +91,36 @@ class BanklyTest extends TestCase
     {
         $this->assertTrue(
             method_exists($this->bankly(), 'getTransferStatus')
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function testEnsureMethodGetTransferFundsExists()
+    {
+        $this->assertTrue(
+            method_exists($this->bankly(), 'getTransferFunds')
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function testEnsureMethodFindTransferFundByAuthCodeExists()
+    {
+        $this->assertTrue(
+            method_exists($this->bankly(), 'findTransferFundByAuthCode')
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function testEnsureMethodGetAccountExists()
+    {
+        $this->assertTrue(
+            method_exists($this->bankly(), 'getAccount')
         );
     }
 }