+++ CHANGELOG.md
@@ -1,7 +1,27 @@
 # Changelog
 
 All notable changes to `bankly-laravel` will be documented in this file
+## 1.4.0 - 2021-03-04
+- Fixed support for PHP8
+- Added Inputs for Customer (Customer, Address, Phone, Analysis)
+- Added Inputs for Documents
 
-## 1.0.0 - 201X-XX-XX
+## 1.3.0 - 2021-01-26
+- Added support to Laravel 8.x
+- Added support to PHP8
+
+## 1.2.0 - 2020-10-22
+- Fixed events params
+- Fixed balance request
+- Added notes on methods that have different responses from staging and production environment
+
+## 1.1.1 - 2020-10-15
+- Fixed /events params
+
+## 1.1.0 - 2020-10-14
+- Added new settings API
+- Removed /baas from staging 
+
+## 1.0.0 - 2020-09-29
 
 - initial release